dial: send wol magic packet multiple times and wait two times Wakeup.Timeout
DIAL spec actually says to "send the packet once every 50ms" because it
is sent "over an unreliable transport layer". it seems to me a bit too
excessive, we send it once at every iteration of the "wakeup" loop.

the spec also recommends "to wait two times the Timeout value" which
seems reasonable.