youtube: added Remote.Add() to add videos to queue
this will be used for the -a flag to add videos without changing what's
currently playing.

the implementation is not very pretty, sometimes the queue gets messed
up and some video gets "lost", I don't know why, I guess I'm doing
something wrong with the api. also we need to send a request for each
video with a delay!