ytcast: allow to specify the interface to use for network operations
the new -i option allows to specify (i.e. restrict) ytcast's network
operations to a single network interface (or ip address or hostname).

this is useful for hosts that have multiple network interfaces and
restrictive firewall rules on some of them since we now allow to not
send M-SEARCH messages etc... to every interface.